+++ main_sweep.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 from prep_instruments import prep_inst
 from gui_input_sweep import get_inputs
 from run_instruments import run_sweep
@@ -24,17 +22,16 @@ try:
         
         for i in range(1, params.dyn_scans+1):
             log.info(f"Start a scan: {i}")
-            scan = run_sweep(pm, laser, params, dryrun=False)
-            scans.append((params.pm_range, scan))            
+            scans.append(run_sweep(pm, laser, params, dryrun=False))            
             params.pm_range -= params.decrement
             
-        data = combine_scans(scans)
         params.pm_range = saved_pm_range
+        data = combine_scans(scans, params)
         
         if not params.reference:
             last_data = data.copy()
         
-        display_plot(data, params=params, ref=last_data, overlays=[e[1] for e in scans])
+        display_plot(data, params=params, ref=last_data, overlays=scans)
             
 except Exception:
     log.exception("Unhandled error")