+++ run_instruments.py
@@ -20,33 +20,37 @@ power_limit = {"10": 10e-3,
              "-60": 1.9999e-9,
              "-70": 199.99e-12}
 
-def run_sweep(pm, laser, params, dryrun=False):
+def run_sweep(pm, laser, params: Params, dryrun=False):
     log.info("--- Running instruments ---")
     
+    power_w_all_ch = []
+    power_dbm_all_ch = []
     tls_wl_start = params.wl_start - params.padding
     tls_wl_stop  = params.wl_stop  + params.padding
 
     # ----- Power Meter -----
-    pm.write(":INIT1:CONT 0")
-    pm.write(":SENS1:FUNC:STAT LOGG, STOP")
-    pm.write(f":SENSE1:CHAN1:POW:WAVE {tls_wl_stop:.3f} NM")
-    pm.write(f":SENSE1:CHAN1:POW:ATIME {params.at_us} US")
-    pm.write(":SENSE1:CHAN1:CORR 0")
-    pm.write(":SENSE1:CHAN1:POW:RANGE:AUTO  0")
-    pm.write(f":SENSE1:CHAN1:POW:RANGE  {params.pm_range} DBM")
-    pm.write(":SENSE1:CHAN1:POW:UNIT  1")   # W (faster)
-    pm.write(":TRIG1:CHAN1:OUTP DIS")
-    pm.write(":TRIG1:CHAN1:INP  CME")
-    pm.write(":TRIG:CONF PASS")
-    pm.write(f":SENSE1:CHAN1:FUNC:PAR:LOGG {params.num_data}, {params.at_us} US")
+    for i in params.channel:
+        pm.write(f":INIT{i}:CONT 0")
+        pm.write(f":SENSE{i}:FUNC:STAT LOGG, STOP")
+        pm.write(f":SENSE{i}:POW:WAVE {tls_wl_stop:.3f} NM")
+        pm.write(f":SENSE{i}:POW:ATIME {params.at_us} US")
+        pm.write(f":SENSE{i}:CORR 0")
+        pm.write(f":SENSE{i}:POW:RANGE:AUTO  0")
+        pm.write(f":SENSE{i}:POW:RANGE  {params.pm_range} DBM")
+        pm.write(f":SENSE{i}:POW:UNIT  1")   # W (faster)
+        pm.write(f":TRIG{i}:OUTP DIS")
+        pm.write(f":TRIG{i}:INP  CME")
+        
+        pm.write(f":SENSE{i}:FUNC:PAR:LOGG {params.num_data}, {params.at_us} US")
+        # PM: arm logging function before sweep starts
+        pm.write(f":SENSE{i}:FUNC:STAT LOGG, START")
 
-    # log.info(f"[PM] Power Range: {pm.query(":SENSE1:CHAN1:POW:RANGE?")}")
+    pm.write(f":TRIG:CONF PASS")
 
     # ----- Laser -----
-        
     laser.write(f":SOURCE0:WAVE  {tls_wl_start:.3f} NM")
-    
     time.sleep(0.1)
+    
     if (w := (float(laser.query(":SOURCE0:WAVE?"))*1e9)) != tls_wl_start:
         log.warning(f"[LASER] The current wavelength: {w}. Laser is still being adjusted.")
         
@@ -63,13 +67,13 @@ def run_sweep(pm, laser, params, dryrun=False):
     laser.write(f":SOURCE0:WAV:SWE:STAR     {tls_wl_start:.3f} NM")
     laser.write(f":SOURCE0:WAV:SWE:STOP     {tls_wl_stop:.3f} NM")
 
+    # ----- Check errors -----
     laser_check_param = (laser.query(":SOUR0:WAV:SWE:CHEC?")).split(',')
     if int(laser_check_param[0]) != 0:
         log.error(f"[LASER] Failed parameter checks: {', '.join(laser_check_param)}")
     else:
         log.info("[LASER] Passed parameter checks")
 
-    # Check system errors
     if (n := int(laser.query(":SYST:ERR:COUN?"))) > 0:
         for _ in range(n):
             log.error(f"[LASER] System error: {laser.query(':SYST:ERR?')}")
@@ -86,9 +90,6 @@ def run_sweep(pm, laser, params, dryrun=False):
 
     if dryrun:
         return None
-    
-    # PM: arm logging function before sweep starts
-    pm.write(":SENS1:FUNC:STAT LOGG, START")
 
     # TLS: starts continuous sweep
     laser.write(":SOURCE0:WAV:SWE:STATE 1")
@@ -98,40 +99,38 @@ def run_sweep(pm, laser, params, dryrun=False):
     while int(laser.query(":SOURCE0:WAV:SWE:STATE?")) == 1:
         log.info("[LASER] Sweeping...")
         time.sleep(1)
-
-    log.info("[LASER] Sweep finished")
-
-    # PM: read logged data
-    log.info("[PM] Read logged measurements")
-    pm.write(":SENS1:FUNC:RES?")
-    time.sleep(2)
-    arr_w = pm.read_binary_values(container=np.ndarray)
-    
-    log.info(f"[PM] Log count: {len(arr_w)}")
-    
-    pm.write(":SENS1:FUNC:STAT LOGG, STOP")
     
     # TLS will stay at the stop wavelength after a sweep
     laser.write(f":SOURCE0:WAVE  {params.wl_stop:.3f} NM")
+    log.info("[LASER] Sweep finished")
 
-    upper_limit = power_limit[str(params.pm_range)]
-    arr_w[arr_w > upper_limit] = np.nan    
-    if np.all(np.isnan(arr_w)):
-        log.error("All measurements are overflown.")
-    elif np.any(np.isnan(arr_w)):
-        log.warning("There are some overflown measurements.")
+    # PM: read logged data
+    for i in params.channel:
+        log.info(f"[PM] Ch.{i}: Read logged measurements")
+        pm.write(f":SENSE{i}:FUNC:RES?")
+        time.sleep(2)
+    
+        power_w_all_ch.append(pm.read_binary_values(container=np.ndarray))
+        log.info(f"[PM] Ch.{i}: Log count: {len(power_w_all_ch[-1])}")
         
-    arr_dbm  = 10 * np.log10(arr_w.astype(np.float64) / 1e-3)  # convert: W -> dBm
-    arr_dbm *= -1                                              # power loss 
+        pm.write(f":SENSE{i}:FUNC:STAT LOGG, STOP")
+        pm.clear()
     
-    # Generate x-axis
-    # wav_stop_tmp = tls_wl_start + (params.step_pm * 1e-3) * (params.num_data - 1)
-    # wav_range    = np.linspace(tls_wl_start, wav_stop_tmp, params.num_data).round(7)
-
-    # i_lo = np.searchsorted(wav_range, params.wl_start - params.step_pm * 1e-3, side='left')
-    # i_hi = np.searchsorted(wav_range, params.wl_stop  + params.step_pm * 1e-3, side='right')
     
-    return arr_dbm
+    upper_limit = power_limit[str(params.pm_range)]
+    
+    for i, arr_w in zip(params.channel, power_w_all_ch):
+        arr_w[arr_w > upper_limit] = np.nan    
+        if np.all(np.isnan(arr_w)):
+            log.warning(f"[PM] Ch.{i}: All measurements are overflown.")
+        elif np.any(np.isnan(arr_w)):
+            log.warning(f"[PM] Ch.{i}: Some measurements are overflown.")
+        
+        arr_dbm  = 10 * np.log10(arr_w.astype(np.float64) / 1e-3)  # convert: W -> dBm
+        arr_dbm *= -1                                              # power loss 
+        power_dbm_all_ch.append(arr_dbm)
+    
+    return power_dbm_all_ch
 
 if __name__ == "__main__":
     pm, laser = prep_inst()